cleaning up

+++ install.sh
@@ -21,16 +21,19 @@ for f in bin/*; do
   ln "$f" "$g"
 done
 
-mkdir -p "$HOME/Library/LaunchAgents"
-for f in LaunchAgents/*; do
-  g="$HOME"/Library/LaunchAgents/$(basename $f);
-  echo "ln $f $g"
-  if [[ "$simulate" == 1 ]]; then
-    continue
-  fi
-  rm -f "$g"
-  ln "$f" "$g"
-done
+files=$(shopt -s nullglob dotglob; echo LaunchAgents/*)
+if (( ${#files} )); then
+  mkdir -p "$HOME/Library/LaunchAgents"
+  for f in LaunchAgents/*; do
+    g="$HOME"/Library/LaunchAgents/$(basename $f);
+    echo "ln $f $g"
+    if [[ "$simulate" == 1 ]]; then
+      continue
+    fi
+    rm -f "$g"
+    ln "$f" "$g"
+  done
+fi
 
 for f in dot.*; do
     g="$HOME/${f/dot/}"
